# Removed auto-decoding from std.algorithm.searching.balancedParens