@wilzbach 2018-01-30
> I'm not sold on this function's usefulness

Me neither, imho it's pretty much useless (or only in very rare occasion useful and shouldn't have been added to Phobos)
BTW a generalization of this - a function that maintains its own stack and returns a range like `untilClosingParentheses` in the ddoc parser on dlang.org would have been useful - see also:
https://github.com/dlang/dlang.org/pull/2060#discussion_r163830331