@andralex 2018-01-30
Yah we should have an algorithm `findMatchingParen` that assumes `r.front` is the opening paren and advances the range until it's positioned on the corresponding closing paren. An optional `rParen` can be passed for unusual closing parens.