@quickfur 2018-01-30
@andralex But to be fully useful, the function will need to count nested parens, so it will need to know what the start parens is. Or are you saying it will just take `r.front` as the opening parens?