@wilzbach 2018-01-30
> Yes, the range is positioned initially on the opening paren, and is advanced through to the corresponding closing paren. The balance of the range is returned (empty if no closing paren).

How would it know what's matching to `[`, `(`, `<`, ... without hard-coding this?