@JackStouffer 2018-01-30
@andralex @quickfur @wilzbach Let's move this discussion to https://issues.dlang.org/show_bug.cgi?id=18336